Enhanced the simulator interface  of 'ARView'
The enhancement of 'ARView' is split into chunks. This change
adds the missing interface mocks for 'DebugOptions', 'Background'
and the missing 'RenderOptions'.
